requirements.txt
Pillow
wand